chore(eslint): consolidate frontend config into root flat config
