+++ eslint.config.mjs
@@ -1,19 +1,60 @@
+import path from 'node:path';
+import { includeIgnoreFile } from '@eslint/compat';
 import js from '@eslint/js';
+import prettier from 'eslint-config-prettier';
+import svelte from 'eslint-plugin-svelte';
+import { defineConfig } from 'eslint/config';
 import globals from 'globals';
-import tseslint from 'typescript-eslint';
+import ts from 'typescript-eslint';
+import svelteConfig from './packages/frontend/svelte.config.js';
 
-export default tseslint.config(
-  {
-    ignores: ['**/dist/**', '**/node_modules/**', '.turbo/**'],
-  },
-  js.configs.recommended,
-  ...tseslint.configs.recommended,
-  {
-    files: ['**/*.ts'],
-    languageOptions: {
-      globals: {
-        ...globals.node,
-      },
-    },
-  }
+const gitignorePath = path.resolve(import.meta.dirname, '.gitignore');
+
+export default defineConfig(
+	includeIgnoreFile(gitignorePath),
+	{ ignores: ['**/dist/**', '**/.svelte-kit/**', '**/.turbo/**', '**/build/**'] },
+	js.configs.recommended,
+	ts.configs.recommended,
+	svelte.configs.recommended,
+	prettier,
+	svelte.configs.prettier,
+	{
+		languageOptions: { globals: { ...globals.browser, ...globals.node } },
+		rules: {
+			// typescript-eslint strongly recommend that you do not use the no-undef lint rule on TypeScript projects.
+			// see: https://typescript-eslint.io/troubleshooting/faqs/eslint/#i-get-errors-from-the-no-undef-rule-about-global-variables-not-being-defined-even-though-there-are-no-typescript-errors
+			'no-undef': 'off'
+		}
+	},
+	{
+		files: ['**/*.svelte', '**/*.svelte.ts', '**/*.svelte.js'],
+		languageOptions: {
+			parserOptions: {
+				projectService: true,
+				extraFileExtensions: ['.svelte'],
+				parser: ts.parser,
+				svelteConfig
+			}
+		}
+	},
+	{
+		// Vendored components — both shadcn-svelte (`ui/`) and Svelte AI Elements
+		// (`ai-elements/`) are regenerated by their respective CLIs/registries.
+		// Do not enforce app-level lint rules on them.
+		files: [
+			'**/components/ui/**/*.{svelte,ts}',
+			'**/components/ai-elements/**/*.{svelte,ts}'
+		],
+		rules: {
+			'svelte/no-navigation-without-resolve': 'off',
+			'svelte/prefer-svelte-reactivity': 'off',
+			'svelte/no-at-html-tags': 'off',
+			'prefer-const': 'off',
+			'no-empty': 'off',
+			'no-useless-assignment': 'off',
+			'@typescript-eslint/no-unused-vars': 'off',
+			'@typescript-eslint/no-explicit-any': 'off',
+			'@typescript-eslint/no-empty-object-type': 'off'
+		}
+	}
 );
