+++ package.json
@@ -27,9 +27,12 @@
     "*.{ts,tsx,js,json,md,yml,yaml}": "prettier --write"
   },
   "devDependencies": {
+    "@eslint/compat": "^2.0.5",
     "@eslint/js": "^10.0.1",
     "@types/bun": "^1.3.6",
     "eslint": "^10.2.1",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-svelte": "^3.17.1",
     "globals": "^17.5.0",
     "lint-staged": "^16.4.0",
     "prettier": "^3.4.2",
